% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Check if required packages are installed}
\usage{
check_packages(output_format)
}
\arguments{
\item{output_format}{Word, Leipzig or Tooltip, desired format}
}
\description{
Calls \code{\link{requireNamespace}} with the required packages.
}
