% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_from_file}
\alias{config_from_file}
\title{Override configuration with a YAML file}
\usage{
config_from_file(filename)
}
\arguments{
\item{filename}{Path to the YAML configuration file, e.g. "glossr-config.yaml".}
}
\value{
Invisibly, the contents of the configuration file that passed validation.
}
\description{
Read configuration from a YAML file to provide instructions for styling
and PDF- or Word-specific options.
}
\examples{
config_file <- system.file("extdata/glossr-config.yml", package="glossr")
config_from_file(config_file)
print_config()
}
