% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss.R
\name{create_gloss}
\alias{create_gloss}
\alias{as_gloss}
\title{Create gloss_data object}
\usage{
create_gloss(
  ...,
  source = NULL,
  translation = NULL,
  label = NULL,
  trans_quotes = getOption("glossr.trans.quotes", "\\"")
)

as_gloss(...)
}
\arguments{
\item{...}{Lines for glossing, or parameters to pass to \code{create_gloss}}

\item{source}{(Optional) Source of example}

\item{translation}{(Optional) Free translation}

\item{label}{(Optional) Example label}

\item{trans_quotes}{(Optional) Quotes to surround the free translation with.}
}
\value{
Object of class `gloss_data`
}
\description{
Based on a character vectors and up to three label arguments,
create an object where those arguments are attributes.
These are:
\describe{
   \item{source}{Where the text comes from. This will be printed
   in the first line of the example, without word alignment.}
   \item{translation}{Free translation. This will be printed as the
   last line of the example, without word alignment and in quotation
   marks if so desired.}
   \item{label}{Named label of the example, for cross-references.}
   \item{lengths}{This is computed within the function, not provider, and
   it's the number of items identified in each gloss line.}
}
}
\section{Functions}{
\itemize{
\item \code{as_gloss}: Render based on R Markdown output
}}

\examples{
ex_sp <- "Un ejemplo en español"
ex_gloss <- "DET.M.SG example in Spanish"
ex_trans <- "An example in Spanish"
my_gloss <- create_gloss(ex_sp, ex_gloss, translation = ex_trans, label="ex1")

# Within R Markdown
as_gloss(my_gloss)
}
