% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elan-annotations.R
\name{elan.annotations}
\alias{elan.annotations}
\title{elan.annotations}
\usage{
elan.annotations(data)
}
\arguments{
\item{data}{the name of the .eaf file to be converted to a dataframe}
}
\value{
a dataframe with the ELAN annotations, participants, start time, and end time
}
\description{
Takes an ELAN .eaf transcription in XML format and saves it as a dataframe
}
\examples{
example <- elan.annotations("data.eaf")

}

