% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_utils.R
\name{format_pdf}
\alias{format_pdf}
\title{Read Latex formatting options}
\usage{
format_pdf(level)
}
\arguments{
\item{level}{Gloss line to format}
}
\value{
Key for expex
}
\description{
Read Latex formatting options
}
