% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_utils.R
\name{format_word}
\alias{format_word}
\alias{format_word_glosses}
\alias{format_word_translation}
\alias{format_word_source}
\title{Read Word formatting options}
\usage{
format_word_glosses(ft)

format_word_translation(ft)

format_word_source(source)
}
\arguments{
\item{ft}{\code{\link[flextable]{flextable}} for gloss lines or translation}

\item{source}{Character vector with the source text.}
}
\value{
Formated table or text
}
\description{
Read Word formatting options
}
\section{Functions}{
\itemize{
\item \code{format_word_glosses}: Format glosses

\item \code{format_word_translation}: Format translation

\item \code{format_word_source}: Format source text
}}

