% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{gloss}
\alias{gloss}
\title{Reference gloss}
\usage{
gloss(label)
}
\arguments{
\item{label}{Label for reference}
}
\value{
Character string with label reference
}
\description{
Latex output uses \code{\@ref(label)} to reference examples,
  whereas HTML output is based on pandoc examples, i.e. \code{(@label)}.
  \code{`r gloss(label)`}, written inline in the text, will return the
  appropriate reference based on the selected output.
}
