% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_utils.R
\name{gloss_format_words}
\alias{gloss_format_words}
\title{Apply latex formatting to many words}
\usage{
gloss_format_words(text, formatting)
}
\arguments{
\item{text}{Character vector of length 1.}

\item{formatting}{Latex formatting code, e.g. \code{textit} or \code{textsc}.}
}
\value{
Reformatted string
}
\description{
Facilitates applying the same latex formatting
  to different words in a row.
}
\examples{
gloss_format_words("Many words to apply italics on.", "textit")
}
