% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{gloss_frame}
\alias{gloss_frame}
\title{Include glosses in R Markdown}
\usage{
gloss_frame(...)
}
\arguments{
\item{...}{One or more glosses.}
}
\value{
Prints text for markdown, in either latex or html formats.
}
\description{
Finishes off the output of \code{\link{gloss_render}} for
  an R Markdown file, based on its output format.
  The input can be one glossed example, multiple ones,
  a list of glossed examples or even a glossed list
  created by \code{\link{gloss_list}}.
}
\details{
This function should be called in an R chunk with
  the `results = "asis"` option (and `echo = FALSE` to avoid
  showing the code). The name of the chunk is
  not relevant.
}
\examples{
ex_sp <- "Un ejemplo en español"
ex_gloss <- "DET.M.SG example in Spanish"
ex_trans <- "An example in Spanish"
my_gloss <- as_gloss(ex_sp, ex_gloss, ex_trans, "ex1")
gloss_frame(my_gloss)
}
