% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{gloss_linesplit}
\alias{gloss_linesplit}
\title{Split lines for HTML}
\usage{
gloss_linesplit(line)
}
\arguments{
\item{line}{Character string to split.}
}
\value{
Character vector of elements.
}
\description{
Splits a character string by spaces keeping groups of words
  surrounded by curly braces together.
}
