% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss_render.R
\encoding{UTF-8}
\name{gloss_render}
\alias{gloss_render}
\alias{gloss_pdf}
\alias{gloss_html}
\alias{gloss_tooltip}
\alias{gloss_leipzig}
\alias{as_gloss}
\alias{gloss_df}
\title{Render a gloss}
\usage{
gloss_pdf(original, parsed, translation = NULL, label = NULL)

gloss_html(original, parsed, translation = NULL, label = NULL)

gloss_tooltip(original, parsed, translation = NULL)

gloss_leipzig(original, parsed, translation = NULL)

as_gloss(original, parsed, translation = NULL, label = NULL)

gloss_df(df)
}
\arguments{
\item{original}{First line for the interlinear gloss,
i.e. example in original language.}

\item{parsed}{Second line for the interlinear gloss,
i.e. glosses for individual words or expression.}

\item{translation}{Free translation (optional).}

\item{label}{Label for the example (optional).}

\item{df}{Dataframe with at least \code{original} and \code{parsed} columns
to generate multiple glosses.}
}
\value{
Object of class \code{gloss}
}
\description{
Render a gloss
}
\section{Functions}{
\itemize{
\item \code{gloss_pdf}: Render in PDF

\item \code{gloss_html}: Render in HTML

\item \code{gloss_tooltip}: Tooltip rendering for HTML

\item \code{gloss_leipzig}: Leipzig.js engine

\item \code{as_gloss}: Render based on R Markdown output

\item \code{gloss_df}: Render glosses from a dataframe
}}

\examples{
ex_sp <- "Un ejemplo en español"
ex_gloss <- "DET.M.SG example in Spanish"
ex_trans <- "An example in Spanish"
gloss_pdf(ex_sp, ex_gloss, ex_trans, "ex1")

gloss_html(ex_sp, ex_gloss, ex_trans, "ex1")

# Within R Markdown
as_gloss(ex_sp, ex_gloss, ex_trans, "ex1")
}
