% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss_render.R
\name{gloss_single}
\alias{gloss_single}
\title{Render a non interlinear gloss}
\usage{
gloss_single(gloss, numbering)
}
\arguments{
\item{gloss}{Object of class \code{\link{gloss_data}}}

\item{numbering}{Whether the gloss should be numbered (in HTML and Word).}
}
\value{
Object of class \code{\link[=new_gloss]{gloss}}.
}
\description{
This function is called when a gloss has only one line beyond the translation
(or even no translation at all). Like other \code{gloss_render()} functions, it
is only meant to be called internally, but it can be used for debugging.
}
\details{
Render a gloss with only one line or one line and free translation in HTML and
Word.
}
