% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{gloss_tooltip}
\alias{gloss_tooltip}
\title{Add tooltip to one word}
\usage{
gloss_tooltip(x, title)
}
\arguments{
\item{x}{Word or expression}

\item{title}{Text of the tooltip}
}
\value{
`shiny.tag` with attributes for a tooltip
}
\description{
Add tooltip to one word
}
\examples{
gloss_tooltip("One", "DET.SG")
}
