% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_utils.R
\name{gloss_word_lines}
\alias{gloss_word_lines}
\title{Process elements of gloss lines for word}
\usage{
gloss_word_lines(gloss_lines)
}
\arguments{
\item{gloss_lines}{Unclassed content of a \code{gloss_data} object}
}
\value{
List of tibbles to print
}
\description{
Process elements of gloss lines for word
}
