% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glosses}
\alias{glosses}
\title{Examples of glosses}
\format{
A \code{tibble} with 5 rows and 6 variables:
\describe{
  \item{original}{The text in the original language.}
  \item{parsed}{The text with translations to English or morphological annotation per word or expression,
  with LaTeX formatting.}
  \item{translation}{Free translation to English.}
  \item{label}{Label for referencing the example.}
  \item{language}{Original language of the text.}
  \item{Source}{Where the example was taken from (published paper).}
}
}
\usage{
glosses
}
\description{
A dataset containing five glossing examples extracted from Koptjevskaja-Tamm (2015)'s
\emph{The Linguistics of Temperature} and chapters within.
}
\keyword{datasets}
