% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossr-package.R
\docType{package}
\name{glossr-package}
\alias{glossr}
\alias{glossr-package}
\title{glossr: Use interlinear glosses in R Markdown}
\description{
Read examples with interlinear glosses from files or from text
and print them in a way compatible with both Latex and HTML outputs.
}
\keyword{internal}
