% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{ignore_latex}
\alias{ignore_latex}
\title{Ignore latex notation}
\usage{
ignore_latex(string)
}
\arguments{
\item{string}{Latex string to purge}
}
\value{
Character vector
}
\description{
Replace most of the latex notation with asterisks.
}
