% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{knit_print.gloss}
\alias{knit_print.gloss}
\title{Print method for glosses}
\usage{
\method{knit_print}{gloss}(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{Other options for knit_print}
}
\description{
Print method for glosses
}
