% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{latex_tag}
\alias{latex_tag}
\title{Regex for a latex tag}
\usage{
latex_tag(tag)
}
\arguments{
\item{tag}{Latex tag}
}
\value{
Regex expression to extract tagged string.
}
\description{
Regex for a latex tag
}
