% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss.R
\name{new_gloss}
\alias{new_gloss}
\title{gloss class}
\usage{
new_gloss(input, output)
}
\arguments{
\item{input}{A \code{\link{gloss_data}} object.}

\item{output}{How the gloss must be printed, depending on the output.}
}
\value{
Object of class \code{gloss}.
}
\description{
The \code{gloss} class contains how a gloss will be printed and its original input
(Object of class \code{\link{gloss_data}}) as \code{data} attribute.
It also has a \code{\link[knitr:knit_print]{knitr::knit_print()}} method for rendering in R Markdown (and Quarto).
}
