% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R, R/utils.R, R/word_utils.R
\name{latex2html}
\alias{latex2html}
\alias{parse_latex}
\alias{latex2word}
\title{Parse latex}
\usage{
latex2html(string)

latex2word(string)
}
\arguments{
\item{string}{Latex string to parse}
}
\value{
Character vector
}
\description{
Parse latex
}
\section{Functions}{
\itemize{
\item \code{latex2html}: Convert to HTML

\item \code{latex2word}: Convert to Word
}}

