% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss-print.R
\name{print_all_words}
\alias{print_all_words}
\title{print_all_words()}
\usage{
print_all_words(data, word)
}
\arguments{
\item{data}{data frame that contains the data from the ELAN file}

\item{word}{word or regular expression that are searching for}
}
\value{
a dataframe of all annotations with the selected search item
}
\description{
Prints annotations that contain a certain regex sequence
}
\examples{
word = "string"
print_all_words(data,word)

}

