% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print_config}
\alias{print_config}
\title{Print the glossr configuration}
\usage{
print_config(config_sections = c("format", "pdf", "word", "other"))
}
\arguments{
\item{config_sections}{Character vector with the following possible values:
\itemize{
\item \strong{format}: to show the formatting options (italics / bold / nothing) for
the different lines.
\item \strong{pdf}: to show $\LaTeX$-specific formatting
\item \strong{word}: to show Word-specific formatting
\item \strong{other}: to show the current output format, translation quotes and
the \code{numbering} setting.
}}
}
\value{
Invisibly, a named list with the printed values
}
\description{
Print the glossr configuration
}
\examples{
print_config()
print_config("pdf") # print only pdf-specific configuration
}
