% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_gloss}
\alias{print_gloss}
\title{Print gloss}
\usage{
print_gloss(sent, format)
}
\arguments{
\item{sent}{Sentence as a list.}

\item{format}{Format to print out with (either \code{html} or \code{latex}).}
}
\description{
The function returns the selected sentence as properly formatted html/latex code.
}
