% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss-print.R
\name{print_selected_time}
\alias{print_selected_time}
\title{print_selected_time()}
\usage{
print_selected_time(data, start_annotation, end_annotation)
}
\arguments{
\item{data}{data frame that contains the data from the ELAN file}

\item{start_annotation}{starting time in seconds to include}

\item{end_annotation}{end time of the annotations to include}
}
\value{
a dataframe with the selected annotations
}
\description{
Prints annotations from a given time
}
\examples{
start_annotation = 120
end_annotation= 200
print_selected_time(data,start_annotation,end_annotation)

}

