% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sentence}
\alias{read_sentence}
\title{Read sentence}
\usage{
read_sentence(path, sentence)
}
\arguments{
\item{path}{The path to the text object as a string.}

\item{sentence}{The name of the sentence object as a string.}
}
\description{
Read sentence file from specified path.
}
