% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_utils.R
\name{reset_max}
\alias{reset_max}
\title{Divide lines for Word}
\usage{
reset_max(m, c, l)
}
\arguments{
\item{m}{Maximum number of characters for a slot}

\item{c}{Cumulative sum of maximums}

\item{l}{Current line}
}
\description{
Helper for \code{\link{gloss_word}}
}
