% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{sc_to_upper}
\alias{sc_to_upper}
\title{Small caps to upper case}
\usage{
sc_to_upper(string)
}
\arguments{
\item{string}{String with a LaTeX small caps tag}
}
\value{
Character vector of length one
}
\description{
Replaces small caps tags from LaTeX to upper case within a string.
}
