% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_default}
\alias{set_default}
\title{Define a default value}
\usage{
set_default(x, default = "")
}
\arguments{
\item{x}{Variable to define}

\item{default}{Default value}
}
\value{
New value
}
\description{
Define a default value
}
