% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{style_options}
\alias{style_options}
\title{List of styling options}
\usage{
style_options(format = c("i", "b"))
}
\arguments{
\item{format}{\code{i} for italics and \code{b} for bold}
}
\value{
Character vector with the ways that a certain format (italics or bold) can be specified.
}
\description{
List of styling options
}
