% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_output}
\alias{validate_output}
\title{Validate output format}
\usage{
validate_output(
  output = c("word", "latex", "leipzig", "tooltip", "html", "pdf")
)
}
\arguments{
\item{output}{Character string with output format required.}
}
\value{
Invisible, the output. It also sets it as the 'glossr.output' option.
}
\description{
Validate output format
}
