% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_utils.R
\name{word_knitr}
\alias{word_knitr}
\title{Apply officer formatting}
\usage{
word_knitr(text, bold = FALSE, italic = FALSE)
}
\arguments{
\item{text}{Text to format}

\item{bold}{Whether the word should be in bold}

\item{italic}{Whether the word should be in italics}
}
\value{
Knitr-ready text
}
\description{
Apply officer formatting
}
