% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss.R
\name{gloss}
\alias{gloss}
\title{gloss}
\usage{
gloss()
}
\arguments{
\item{data}{data frame that contains examples to be glossed}

\item{option1}{first option to be defined later}

\item{option2}{second option to be defined later}
}
\value{
a gloss
}
\description{
Prints a gloss
}
\examples{
gloss()

}

