% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss_render.R
\name{gloss_df}
\alias{gloss_df}
\title{Render gloss from a dataframe}
\usage{
gloss_df(df)
}
\arguments{
\item{df}{Dataframe one row per gloss. Columns \code{translation},
\code{source} and \code{label} have special meaning
(see \code{\link{new_gloss_data}}); all the others will be interpreted as
lines to align in the order given.}
}
\value{
Object of class \code{gloss} with the original input as \code{data} attribute.
}
\description{
Render gloss from a dataframe
}
\examples{
my_gloss <- data.frame(
  first_line = "my first line",
  second_line = "my second line",
  translation = "Translation of my example",
  label = "label"
)
gloss_df(my_gloss)
}
