% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\encoding{UTF-8}
\name{gloss_linetooltip}
\alias{gloss_linetooltip}
\title{Apply tooltip to a full gloss}
\usage{
gloss_linetooltip(original, parsed)
}
\arguments{
\item{original}{Text to show in the \code{tooltip} rendering.}

\item{parsed}{Text to show as tooltip when hovering}
}
\value{
List of `shiny.tag`
}
\description{
Apply tooltip to a full gloss
}
\examples{
ex_sp <- "Un ejemplo en español"
ex_gloss <- "DET.M.SG example in Spanish"
gloss_linetooltip(ex_sp, ex_gloss)
}
