% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_utils.R
\name{gloss_list}
\alias{gloss_list}
\title{Sublist glosses}
\usage{
gloss_list(..., listlabel = NULL)
}
\arguments{
\item{...}{Glosses}

\item{listlabel}{Label for the full list (optional)}
}
\value{
Character vector including the frame for a list of glosses.
}
\description{
Takes a series of glosses from \code{\link{gloss_render}}
  and puts them in a list within one example for PDF output.
}
