% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss_render.R
\encoding{UTF-8}
\name{gloss_render}
\alias{gloss_render}
\alias{gloss_pdf}
\alias{gloss_html}
\alias{gloss_tooltip}
\alias{gloss_leipzig}
\alias{gloss_df}
\title{Render a gloss}
\usage{
gloss_pdf(gloss)

gloss_html(gloss)

gloss_tooltip(gloss)

gloss_leipzig(gloss)

gloss_df(df)
}
\arguments{
\item{gloss}{Object of class \code{gloss_data}}

\item{df}{Dataframe one row per gloss. Columns \code{translation},
\code{source} and \code{label} have special meaning
(see \code{\link{create_gloss}}); all the others will be interpreted as
lines to align in the order given.}
}
\value{
Object of class \code{gloss}
}
\description{
Render a gloss
}
\section{Functions}{
\itemize{
\item \code{gloss_pdf}: Render in PDF

\item \code{gloss_html}: Render in HTML

\item \code{gloss_tooltip}: Tooltip rendering for HTML

\item \code{gloss_leipzig}: Leipzig.js engine

\item \code{gloss_df}: Render glosses from a dataframe
}}

\examples{
ex_sp <- "Un ejemplo en español"
ex_gloss <- "DET.M.SG example in Spanish"
ex_trans <- "An example in Spanish"
my_gloss <- create_gloss(ex_sp, ex_gloss, translation = ex_trans, label="ex1")
gloss_pdf(my_gloss)

gloss_html(my_gloss)
}
