% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss_render.R
\encoding{UTF-8}
\name{gloss_render}
\alias{gloss_render}
\alias{gloss_pdf}
\alias{gloss_html}
\title{Render a gloss}
\usage{
gloss_pdf(original, parsed, translation = NULL, label = NULL)

gloss_html(original, parsed, translation = NULL, label = NULL)
}
\arguments{
\item{original}{First line for the interlinear gloss,
i.e. example in original language.}

\item{parsed}{Second line for the interlinear gloss,
i.e. glosses for individual words or expression.}

\item{translation}{Free translation (optional).}

\item{label}{Label for the example (optional).}
}
\value{
Character vector to be printed.
}
\description{
Render a gloss
}
\section{Functions}{
\itemize{
\item \code{gloss_pdf}: Render in PDF

\item \code{gloss_html}: Render in HTML
}}

\examples{
ex_sp <- "Un ejemplo en español"
ex_gloss <- "DET.M.SG example in Spanish"
ex_trans <- "An example in Spanish"
gloss_pdf(ex_sp, ex_gloss, ex_trans, "ex1")

gloss_html(ex_sp, ex_gloss, ex_trans, "ex1")
}
