% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_utils.R
\name{gloss_table}
\alias{gloss_table}
\title{Create table from a gloss}
\usage{
gloss_table(gloss_output, is_translation = FALSE)
}
\arguments{
\item{gloss_output}{Gloss lines as a table for Word}

\item{is_translation}{Whether the table is for the free translation line.}
}
\value{
\code{\link[flextable]{flextable}} object
}
\description{
Create table from a gloss
}
