% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossr-package.R
\docType{package}
\name{glossr-package}
\alias{glossr}
\alias{glossr-package}
\title{glossr: Use interlinear glosses in R Markdown}
\description{
Read examples with interlinear glosses from files or from text
and print them in a way compatible with both Latex and HTML outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://montesmariana.github.io/glossr/}
  \item \url{https://github.com/montesmariana/glossr}
}

}
\author{
\strong{Maintainer}: Mariana Montes \email{montesmariana@gmail.com} (\href{https://orcid.org/0000-0002-3869-3207}{ORCID})

Other contributors:
\itemize{
  \item Benjamin Chauvette (Author of included leipzig.js library) [copyright holder]
}

}
\keyword{internal}
