% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{use_glossr}
\alias{use_glossr}
\title{Use glossr}
\usage{
use_glossr(output_format = NULL, styling = list())
}
\arguments{
\item{output_format}{Desired output format}

\item{styling}{Named list of styling options for specific elements of glosses.}
}
\value{
Set options
}
\description{
Override default configuration.
}
\examples{
use_glossr(styling = list(numbering = FALSE, trans_quotes = "~"))
print_config("other")

}
\seealso{
\code{\link[=set_style_options]{set_style_options()}}
}
