% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossr-package.R
\name{use_glossr}
\alias{use_glossr}
\title{Use glossr}
\usage{
use_glossr(html_format = NULL)
}
\arguments{
\item{html_format}{Whether the html output should use leipzig.js or tooltips.}
}
\value{
Set glossr.output option
}
\description{
Call in a setup chunk.
}
